SLM-147 Split resource tests as more are coming
